Fix fact oneadmin_pubkey_rsa so if the user has no key, it does not get set.

+++ lib/facter/oneadmin_pubkey_rsa.rb
@@ -1,10 +1,14 @@
 require 'etc'
 
-Facter.add(:oneadmin_pubkey_rsa) do
-  setcode do
-    ent = Etc.getpwnam("oneadmin")
-    keyfile = File.open(ent.dir + "/.ssh/id_rsa.pub")
-    keyarray = keyfile.read.split(" ")
-    keyarray[1]
+ent = Etc.getpwnam("oneadmin")
+if File.exists?(ent.dir + "/.ssh/id_rsa.pub") then
+
+  Facter.add(:oneadmin_pubkey_rsa) do
+    setcode do
+      keyfile = File.open(ent.dir + "/.ssh/id_rsa.pub")
+      keyarray = keyfile.read.split(" ")
+      keyarray[1]
+    end
   end
+
 end
